// selenium/utils/types.ts
export enum NAVIGATION_ITEMS {
    CATALOG = "https://catalog.onliner.by",
    AUTO = "https://ab.onliner.by",
    REALT = "https://r.onliner.by/pk",
    TASKS = "https://s.onliner.by/tasks",
    BARAHOLKA = "https://baraholka.onliner.by/",
}

export enum SELECTOR_TYPES {
    CSS = 'css',
    XPATH = 'xpath',
    CLASS_NAME = 'className',
    ID = 'id',
}

export enum PAGES {
    HOME = "HomePage",
    LOGIN = "LoginPage",
}